COM-1960 - update tests
